Refactor multiple backstacks implementation
This commit refactors the multiple backstacks recipe by extracting core navigation logic into separate `Navigator` and `NavigationState` classes.

- `NavigationState` is now a dedicated state holder for the current top-level route and the back stacks for each tab.
- `Navigator` encapsulates the navigation logic for handling forward and back navigation events.
- The `MultipleStacksActivity` is simplified to use these new classes.
- The `RouteB1` destination has been simplified from a data class to an object, and click handlers have been updated accordingly.
